Create task-30.md (#1)
The example in English mentions years, which in Spanish is written as "año". The letter "ñ" is an accented character that doesn't exist in the ASCII table. Since input data is supposed to be comprised of only latin letters, the example might need to be changed; "años" could be replaced with "meses" (which means "months").